Squashed 'airdcpp-core/' changes from d86d1f01d..e2a226a4a
e2a226a4a Fix a deadlock
b9bf6fb88 Cleanup

git-subtree-dir: airdcpp-core
git-subtree-split: e2a226a4a3da28386fa3e65da11900974c42d15e

+++ airdcpp/StartupParams.h
@@ -27,7 +27,7 @@ class StartupParams {
 public:
 	typedef deque<string> ParamList;
 
-	bool hasParam(const string& aParam, size_t aPos = -1) const noexcept;
+	bool hasParam(const string& aParam, int aPos = -1) const noexcept;
 	void addParam(const string& aParam) noexcept;
 	bool removeParam(const string& aParam) noexcept;
 	optional<string> getValue(const string& aKey) const noexcept;
